+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/api/SkinType.java
@@ -1,0 +1,5 @@
+package io.github.scroojalix.npcmanager.utils.api;
+
+public enum SkinType {
+    USERNAME, URL;
+}