+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/npc/skin/SkinType.java
@@ -1,5 +1,0 @@
-package io.github.scroojalix.npcmanager.utils.npc.skin;
-
-public enum SkinType {
-    USERNAME, URL;
-}